デバッグコードの消去とソースコードの整理

+++ ACG/ACG/ObjSign.cpp
@@ -15,8 +15,8 @@ CObjSign::CObjSign(int x, int y)
 {
 	m_map_x = x;
 	m_map_y = y;
-	m_px = x * BLOCK_SIZE;
-	m_py = y * BLOCK_SIZE;
+	m_px =(float)x * BLOCK_SIZE;
+	m_py =(float)y * BLOCK_SIZE;
 }
 
 //イニシャライズ
